django_bot/requirements.txt
aiogram==3.4.1
asgiref==3.7.2
astroid==2.15.7
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.7
Django==4.2.5
django-base64field==1.0
django-colorfield==0.10.1
django-cors-headers==3.13.0
django-filter==23.3
django-import-export==3.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
et-xmlfile==1.1.0
flake8==6.1.0
gunicorn==20.1.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
MarkupPy==1.14
mccabe==0.7.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
platformdirs==3.10.0
psycopg2-binary==2.9.3
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pylint==2.17.6
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
tablib==3.5.0
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
async-timeout==4.0.3
emoji==2.11.0